Update Calculations.h
Changing hypot() function data types from double to float

+++ CIS227-cpp-Group4-Assignment1/Calculations.h
@@ -10,7 +10,7 @@
 using namespace std;
 
 //declare functions
-double hypot(double a, double b);
+float hypot(float a, float b);
 
 float area(float a, float b, float h);
